Move Pid type into pid module
Move the Pid type into the pid module, where it is easier to extend it.
