+++ src/pid.rs
@@ -1,0 +1,29 @@
+use std::fmt::Display;
+use std::fmt::Formatter;
+use std::fmt::Result as FmtResult;
+use std::num::NonZeroU32;
+
+
+/// An enumeration identifying a process.
+#[derive(Clone, Copy, Debug)]
+pub enum Pid {
+    /// The current process.
+    Slf,
+    /// The process identified by the provided ID.
+    Pid(NonZeroU32),
+}
+
+impl Display for Pid {
+    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
+        match self {
+            Self::Slf => write!(f, "self"),
+            Self::Pid(pid) => write!(f, "{pid}"),
+        }
+    }
+}
+
+impl From<u32> for Pid {
+    fn from(pid: u32) -> Self {
+        NonZeroU32::new(pid).map(Pid::Pid).unwrap_or(Pid::Slf)
+    }
+}
